improvements to generic components
