+++ Maestri.Front/app/src/components/backgroud/styles.js
