+++ Maestri.Front/app/src/components/information/index.js
@@ -1,0 +1,14 @@
+import { Text, View } from "react-native"
+import stylesComponent from "../../design-System/components-styles.js"
+
+const Information = () => {
+    return (
+        <View style={stylesComponent.container}>
+            <Text> um simples teste lol
+                 </Text>
+        </View>
+    )
+}
+
+
+export default Information
