+++ Maestri.Front/app/src/pages/complementaryRegister/complementaryRegister-II.js
@@ -1,0 +1,16 @@
+import { View,TouchableOpacity,Text } from "react-native"
+
+import stylesComponent from "../../design-System/components-styles";
+import Form from "../../components/form";
+import Title from "../../components/title";
+
+
+const ComplementaryRegisterII  = ()=>{
+    return(
+        <View>
+
+        </View>
+    )
+}
+
+export default ComplementaryRegisterII
