+++ Maestri.Front/app/src/pages/complementaryRegister/complementaryRegister-III.js
@@ -1,0 +1,16 @@
+import { View,TouchableOpacity,Text } from "react-native"
+
+import stylesComponent from "../../design-System/components-styles";
+import Form from "../../components/form";
+import Title from "../../components/title";
+
+
+const ComplementaryRegisterIII = ()=>{
+    return(
+        <View>
+
+        </View>
+    )
+}
+
+export default ComplementaryRegisterIII
