+++ Maestri.Front/app/src/pages/complementaryRegister/formState.js
@@ -1,0 +1,17 @@
+
+
+class FormComplementionRegisterState {
+    #formInputsFields = {};
+
+    setValues({...inputs}){
+        console.log(inputs);
+    }
+
+    getValues(){
+        return this.#formInputsFields;
+    }
+}
+
+const formComplementionRegisterState = new FormComplementionRegisterState();
+
+export default formComplementionRegisterState
