+++ Maestri.Front/app/src/pages/resources/index.js
@@ -1,0 +1,10 @@
+import { Text, View } from "react-native"
+import Information from "../../components/information"
+
+const Resources = ()=>{
+    return (
+        <Information/>
+    )
+}
+
+export default Resources
