+++ Maestri.Front/package.json
@@ -18,7 +18,8 @@
     "react-native": "0.72.6",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "safe-regex": "^2.1.1"
+    "safe-regex": "^2.1.1",
+    "@react-native-community/datetimepicker": "7.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
